+++ lib/music_xml.dart
@@ -1,5 +1,8 @@
+export 'src/basic_attributes.dart';
+export 'src/barline.dart';
 export 'src/chord_symbol.dart';
 export 'src/key_signature.dart';
+export 'src/kind.dart';
 export 'src/lyric.dart';
 export 'src/measure.dart';
 export 'src/music_xml_document.dart';
@@ -7,7 +10,9 @@ export 'src/music_xml_parser_state.dart';
 export 'src/note.dart';
 export 'src/note_duration.dart';
 export 'src/part.dart';
+export 'src/pitch.dart';
 export 'src/print.dart';
+export 'src/root.dart';
 export 'src/score_part.dart';
 export 'src/tempo.dart';
 export 'src/time_signature.dart';