+++ test/music_xml_test.dart
@@ -1,9 +1,6 @@
 import 'dart:io';
 
 import 'package:music_xml/music_xml.dart';
-import 'package:music_xml/src/barline.dart';
-import 'package:music_xml/src/kind.dart';
-import 'package:music_xml/src/basic_attributes.dart';
 import 'package:test/test.dart';
 import 'package:xml/xml.dart';
 
@@ -12,7 +9,7 @@ final file = File('test/assets/musicXML.xml');
 void main() {
   group('constructor', () {
     test('ChordSymbol', () {
-      expect(ChordSymbol(), isNotNull);
+      expect(ChordSymbol.noChord, isNotNull);
     });
     test('KeySignature', () {
       expect(KeySignature(), isNotNull);
@@ -24,15 +21,15 @@ void main() {
       expect(ScorePart(), isNotNull);
     });
     test('Measure', () {
-      expect(Measure(), isNotNull);
+      expect(Measure(number: 0), isNotNull);
     });
     test('NoteDuration', () {
       expect(
           NoteDuration(
-            null,
-            null,
-            null,
-            null,
+            1,
+            1,
+            1.0,
+            0,
             0,
             '',
             0.0,
@@ -51,10 +48,10 @@ void main() {
             false,
             false,
             NoteDuration(
-              null,
-              null,
-              null,
-              null,
+              0,
+              0,
+              0,
+              0,
               0,
               '',
               0.0,
@@ -63,25 +60,37 @@ void main() {
             null,
             null,
             null,
+            [],
           ),
           isNotNull);
     });
     test('Part', () {
-      expect(Part('id', ScorePart(), [Measure()]), isNotNull);
+      expect(Part('id', ScorePart(), [Measure(number: 0)]), isNotNull);
     });
     test('Tempo', () {
       expect(Tempo(0, 0), isNotNull);
     });
     test('TimeSignature', () {
-      expect(TimeSignature(), isNotNull);
+      final timeSignature = TimeSignature(
+        divisions: 2,
+        numerator: 4,
+        denominator: 8,
+        timePosition: 0,
+      );
+
+      expect(timeSignature.beats, 2);
+      expect(timeSignature.beatType, 4);
+      expect(timeSignature.numerator, 4);
+      expect(timeSignature.denominator, 8);
     });
     test('MusicXmlDocument', () {
       expect(
         MusicXmlDocument(
+          'Title',
           XmlDocument([]),
           {'scorePart': ScorePart()},
           [
-            Part('id', ScorePart(), [Measure()]),
+            Part('id', ScorePart(), [Measure(number: 0)]),
           ],
           0.0,
         ),
@@ -101,6 +110,11 @@ void main() {
       expect(document.totalTimeSecs, closeTo(49.5, 1E-1));
     });
 
+    test('title', () {
+      final title = document.title;
+      expect(title, 'It\'s All In The Game');
+    });
+
     test('ScorePart.parse', () {
       final scorePart = document.scoreParts.values.single;
       expect(scorePart.id, 'P1');
@@ -118,6 +132,7 @@ void main() {
 
     test('Measure.parse', () {
       final measure = document.parts.single.measures.first;
+      expect(measure.number, 1);
       expect(measure.notes.length, 7);
       expect(measure.chordSymbols.length, 0);
       expect(measure.tempos.length, 0);
@@ -150,6 +165,14 @@ void main() {
       expect(note.pitch?.value, 70);
     });
 
+    test('Pitch.parse', () {
+      final pitch =
+          document.parts.single.measures.first.notes.last.pitchTypeSafe!;
+      expect(pitch.step, Step.b);
+      expect(pitch.octave, 4);
+      expect(pitch.alter, -1.0);
+    });
+
     group('Lyric.parse', () {
       test('with a note containing multiple lyrics', () {
         final note3 = document.parts.single.measures.first.notes[2];
@@ -178,9 +201,54 @@ void main() {
     });
 
     test('Tie.parse', () {
+      // Get two notes that are tied together
       final measures = document.parts.single.measures;
-      expect(measures[7].notes.first.tie?.type, StartStop.start);
-      expect(measures[8].notes.first.tie?.type, StartStop.stop);
+      final startNote = measures[7].notes.first;
+      final stopNote = measures[8].notes.first;
+      final anotherNote = measures[6].notes.first;
+      expect(startNote.ties.first.type, StartStop.start);
+      expect(stopNote.ties.first.type, StartStop.stop);
+
+      // Tied notes should have the same noteId
+      expect(startNote.noteId != 0, isTrue);
+      expect(startNote.noteId, stopNote.noteId);
+      expect(anotherNote.noteId, isNot(startNote.noteId));
+
+      // Check isNoteOn, isNoteOff
+      expect(startNote.isNoteOn, isTrue);
+      expect(startNote.isNoteOff, isFalse);
+
+      expect(stopNote.isNoteOn, isFalse);
+      expect(stopNote.isNoteOff, isTrue);
+
+      // Check continuesOtherNote and isContinuedByOtherNote
+      expect(startNote.continuesOtherNote, isFalse);
+      expect(startNote.isContinuedByOtherNote, isTrue);
+
+      expect(stopNote.continuesOtherNote, isTrue);
+      expect(stopNote.isContinuedByOtherNote, isFalse);
+
+      // Tied duration should be the sum of the durations of tied notes
+      expect(
+        startNote.noteDurationTied.seconds,
+        startNote.noteDuration.seconds + stopNote.noteDuration.seconds,
+      );
+
+      expect(
+        stopNote.noteDurationTied.seconds,
+        startNote.noteDuration.seconds + stopNote.noteDuration.seconds,
+      );
+
+      // Time position should be the position of the first note
+      expect(
+        startNote.noteDurationTied.timePosition,
+        startNote.noteDuration.timePosition,
+      );
+
+      expect(
+        stopNote.noteDurationTied.timePosition,
+        startNote.noteDuration.timePosition,
+      );
     });
 
     test('Duration.parse', () {
@@ -207,8 +275,8 @@ void main() {
 
     test('Root.parse', () {
       final chordSymbol = document.parts.single.measures[2].chordSymbols.first;
-      expect(chordSymbol.rootTypeSafe?.alter, -1);
-      expect(chordSymbol.rootTypeSafe?.step, Step.b);
+      expect(chordSymbol.rootTypeSafe.alter, -1);
+      expect(chordSymbol.rootTypeSafe.step, Step.b);
       expect(chordSymbol.kindTypeSafe, Kind.major);
     });
 
@@ -226,6 +294,43 @@ void main() {
       expect(keySignature!.key, -1);
       expect(keySignature.mode, 'major');
       expect(keySignature.timePosition, 0);
+    });
+
+    test('Kind', () {
+      expect(Kind.undefined.simple, SimpleKind.other);
+      expect(Kind.augmented.simple, SimpleKind.augmented);
+      expect(Kind.augmentedSeventh.simple, SimpleKind.augmented);
+      expect(Kind.diminished.simple, SimpleKind.diminished);
+      expect(Kind.diminishedSeventh.simple, SimpleKind.diminished);
+      expect(Kind.dominant.simple, SimpleKind.major);
+      expect(Kind.dominant11th.simple, SimpleKind.major);
+      expect(Kind.dominant13th.simple, SimpleKind.major);
+      expect(Kind.dominantNinth.simple, SimpleKind.major);
+      expect(Kind.trench.simple, SimpleKind.other);
+      expect(Kind.german.simple, SimpleKind.other);
+      expect(Kind.halfDiminished.simple, SimpleKind.diminished);
+      expect(Kind.italian.simple, SimpleKind.other);
+      expect(Kind.major.simple, SimpleKind.major);
+      expect(Kind.major11th.simple, SimpleKind.major);
+      expect(Kind.major13th.simple, SimpleKind.major);
+      expect(Kind.majorMinor.simple, SimpleKind.major);
+      expect(Kind.majorNinth.simple, SimpleKind.major);
+      expect(Kind.majorSeventh.simple, SimpleKind.major);
+      expect(Kind.majorSixth.simple, SimpleKind.major);
+      expect(Kind.minor.simple, SimpleKind.minor);
+      expect(Kind.minor11th.simple, SimpleKind.minor);
+      expect(Kind.minor13th.simple, SimpleKind.minor);
+      expect(Kind.minorNinth.simple, SimpleKind.minor);
+      expect(Kind.minorSeventh.simple, SimpleKind.minor);
+      expect(Kind.minorSixth.simple, SimpleKind.minor);
+      expect(Kind.neapolitan.simple, SimpleKind.major);
+      expect(Kind.none.simple, SimpleKind.other);
+      expect(Kind.other.simple, SimpleKind.other);
+      expect(Kind.pedal.simple, SimpleKind.other);
+      expect(Kind.power.simple, SimpleKind.other);
+      expect(Kind.suspendedFourth.simple, SimpleKind.sus);
+      expect(Kind.suspendedSecond.simple, SimpleKind.sus);
+      expect(Kind.tristan.simple, SimpleKind.other);
     });
   });
 }